expansion-vis/backend/requirements.txt
brotlipy==0.7.0
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
conda==4.10.3
contourpy==1.1.1
cycler==0.12.1
Flask==2.2.0
Flask-Cors==4.0.0
Flask-Script==2.0.6
flask-session==0.5.0
fonttools==4.53.0
importlib-metadata==6.8.0
importlib-resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
pillow==10.2.0
pycosat==0.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scipy==1.10.1
threadpoolctl==3.5.0
# torch==1.12.1+cu113
# torchaudio==0.12.1+cu113
# torchvision==0.13.1+cu113
tzdata==2024.1
Werkzeug==2.2.2
zipp==3.17.0
